requirements.txt
google-generativeai>=0.8.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
rich>=13.0.0
python-dotenv>=1.0.0
